+++ tools/node.js
@@ -1,5 +1,5 @@
 ////// Manual Node.js Testing
 
-global.rtv = require('../dist/rtv.umd');
+global.rtv = require('../dist/umd/rtv.dev');
 global.ld = require('lodash');
 global.ostr = function(v) { return Object.prototype.toString.call(v); };